Reorganize lib/projects
