+++ commands/app/migrate.ts
@@ -18,7 +18,7 @@ import { EXIT_CODES } from '../../lib/enums/exitCodes';
 import { migrateApp2025_2, MigrateAppArgs } from '../../lib/app/migrate';
 import { uiBetaTag, uiCommandReference, uiLink } from '../../lib/ui';
 import { migrateApp2023_2 } from '../../lib/app/migrate_legacy';
-import { getIsInProject } from '../../lib/projects';
+import { getIsInProject } from '../../lib/projects/config';
 
 const { v2023_2, v2025_2, unstable } = PLATFORM_VERSIONS;
 export const validMigrationTargets = [v2023_2, v2025_2, unstable];
