+++ lib/projects/ensureProjectExists.ts
@@ -1,132 +1,21 @@
-import fs from 'fs-extra';
-import path from 'path';
-import findup from 'findup-sync';
 import { logger } from '@hubspot/local-dev-lib/logger';
 import {
   createProject,
   fetchProject,
 } from '@hubspot/local-dev-lib/api/projects';
 import { isSpecifiedError } from '@hubspot/local-dev-lib/errors/index';
-import { getCwd, getAbsoluteFilePath } from '@hubspot/local-dev-lib/path';
 import { Project } from '@hubspot/local-dev-lib/types/Project';
 import { HubSpotPromise } from '@hubspot/local-dev-lib/types/Http';
 
-import {
-  FEEDBACK_INTERVAL,
-  DEFAULT_POLLING_DELAY,
-  PROJECT_CONFIG_FILE,
-} from '../constants';
+import { DEFAULT_POLLING_DELAY } from '../constants';
 import { promptUser } from '../prompts/promptUtils';
 import { EXIT_CODES } from '../enums/exitCodes';
-import { uiLine, uiAccountDescription, uiCommandReference } from '../ui';
+import { uiAccountDescription } from '../ui';
 import { i18n } from '../lang';
 import SpinniesManager from '../ui/SpinniesManager';
-import { ProjectConfig } from '../../types/Projects';
 import { logError, ApiErrorContext } from '../errorHandlers/index';
 
 const i18nKey = 'lib.projects';
-
-export function writeProjectConfig(
-  configPath: string,
-  config: ProjectConfig
-): boolean {
-  try {
-    fs.ensureFileSync(configPath);
-    fs.writeFileSync(configPath, JSON.stringify(config, null, 2));
-    logger.debug(`Wrote project config at ${configPath}`);
-  } catch (e) {
-    logger.debug(e);
-    return false;
-  }
-  return true;
-}
-
-export function getIsInProject(dir?: string): boolean {
-  const configPath = getProjectConfigPath(dir);
-  return !!configPath;
-}
-
-function getProjectConfigPath(dir?: string): string | null {
-  const projectDir = dir ? getAbsoluteFilePath(dir) : getCwd();
-
-  const configPath = findup(PROJECT_CONFIG_FILE, {
-    cwd: projectDir,
-    nocase: true,
-  });
-
-  return configPath;
-}
-
-export interface LoadedProjectConfig {
-  projectDir: string | null;
-  projectConfig: ProjectConfig | null;
-}
-
-export async function getProjectConfig(
-  dir?: string
-): Promise<LoadedProjectConfig> {
-  const configPath = getProjectConfigPath(dir);
-  if (!configPath) {
-    return { projectConfig: null, projectDir: null };
-  }
-
-  try {
-    const config = fs.readFileSync(configPath);
-    const projectConfig: ProjectConfig = JSON.parse(config.toString());
-    return {
-      projectDir: path.dirname(configPath),
-      projectConfig,
-    };
-  } catch (e) {
-    logger.error('Could not read from project config');
-    return { projectConfig: null, projectDir: null };
-  }
-}
-
-export function validateProjectConfig(
-  projectConfig: ProjectConfig | null,
-  projectDir: string | null
-): asserts projectConfig is ProjectConfig {
-  if (!projectConfig || !projectDir) {
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.configNotFound`, {
-        createCommand: uiCommandReference('hs project create'),
-      })
-    );
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  if (!projectConfig.name || !projectConfig.srcDir) {
-    logger.error(i18n(`${i18nKey}.validateProjectConfig.configMissingFields`));
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  const resolvedPath = path.resolve(projectDir, projectConfig.srcDir);
-  if (!resolvedPath.startsWith(projectDir)) {
-    const projectConfigFile = path.relative(
-      '.',
-      path.join(projectDir, PROJECT_CONFIG_FILE)
-    );
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.srcOutsideProjectDir`, {
-        srcDir: projectConfig.srcDir,
-        projectConfig: projectConfigFile,
-      })
-    );
-    return process.exit(EXIT_CODES.ERROR);
-  }
-
-  if (!fs.existsSync(resolvedPath)) {
-    logger.error(
-      i18n(`${i18nKey}.validateProjectConfig.srcDirNotFound`, {
-        srcDir: projectConfig.srcDir,
-        projectDir: projectDir,
-      })
-    );
-
-    return process.exit(EXIT_CODES.ERROR);
-  }
-}
 
 async function pollFetchProject(
   accountId: number,
@@ -237,14 +126,5 @@ export async function ensureProjectExists(
     }
     logError(err, new ApiErrorContext({ accountId }));
     process.exit(EXIT_CODES.ERROR);
-  }
-}
-
-export function logFeedbackMessage(buildId: number): void {
-  if (buildId > 0 && buildId % FEEDBACK_INTERVAL === 0) {
-    uiLine();
-    logger.log(i18n(`${i18nKey}.logFeedbackMessage.feedbackHeader`));
-    uiLine();
-    logger.log(i18n(`${i18nKey}.logFeedbackMessage.feedbackMessage`));
   }
 }
