+++ lib/projects/localDev/helpers.ts
@@ -17,52 +17,58 @@ import { CLIAccount } from '@hubspot/local-dev-lib/types/Accounts';
 import { Environment } from '@hubspot/local-dev-lib/types/Config';
 import { DeveloperTestAccount } from '@hubspot/local-dev-lib/types/developerTestAccounts';
 import { Project } from '@hubspot/local-dev-lib/types/Project';
+import { Build } from '@hubspot/local-dev-lib/types/Build';
+import { getSandboxUsageLimits } from '@hubspot/local-dev-lib/api/sandboxHubs';
+
+import {
+  ProjectConfig,
+  ProjectPollResult,
+  ProjectSubtask,
+} from '../../../types/Projects';
+import { ProjectDevTargetAccountPromptResponse } from '../../../types/Prompts';
 
 import {
   confirmDefaultAccountPrompt,
   selectSandboxTargetAccountPrompt,
   selectDeveloperTestTargetAccountPrompt,
   confirmUseExistingDeveloperTestAccountPrompt,
-} from './prompts/projectDevTargetAccountPrompt';
-import { confirmPrompt } from './prompts/promptUtils';
-import { validateSandboxUsageLimits, getAvailableSyncTypes } from './sandboxes';
-import { syncSandbox } from './sandboxSync';
-import { validateDevTestAccountUsageLimits } from './developerTestAccounts';
-import { uiCommandReference, uiLine, uiAccountDescription } from './ui';
-import SpinniesManager from './ui/SpinniesManager';
-import { i18n } from './lang';
-import { EXIT_CODES } from './enums/exitCodes';
-import { trackCommandMetadataUsage } from './usageTracking';
+} from '../../prompts/projectDevTargetAccountPrompt';
+import { confirmPrompt } from '../../prompts/promptUtils';
+import {
+  validateSandboxUsageLimits,
+  getAvailableSyncTypes,
+} from '../../sandboxes';
+import { syncSandbox } from '../../sandboxSync';
+import { validateDevTestAccountUsageLimits } from '../../developerTestAccounts';
+import { uiCommandReference, uiLine, uiAccountDescription } from '../../ui';
+import SpinniesManager from '../../ui/SpinniesManager';
+import { EXIT_CODES } from '../../enums/exitCodes';
+import { trackCommandMetadataUsage } from '../../usageTracking';
 import {
   isAppDeveloperAccount,
   isDeveloperTestAccount,
   isUnifiedAccount,
-} from './accountTypes';
-import { handleProjectUpload } from './projects/upload';
-import { pollProjectBuildAndDeploy } from './projects/buildAndDeploy';
+} from '../../accountTypes';
+import { handleProjectUpload } from '../../projects/upload';
+import { pollProjectBuildAndDeploy } from '../../projects/buildAndDeploy';
 import {
   PROJECT_ERROR_TYPES,
   PROJECT_BUILD_TEXT,
   PROJECT_DEPLOY_TEXT,
-} from './constants';
-import { logError, ApiErrorContext, debugError } from './errorHandlers/index';
+} from '../../constants';
+import {
+  logError,
+  ApiErrorContext,
+  debugError,
+} from '../../errorHandlers/index';
 import {
   buildSandbox,
   buildDeveloperTestAccount,
   saveAccountToConfig,
-} from './buildAccount';
-import { hubspotAccountNamePrompt } from './prompts/accountNamePrompt';
-import {
-  ProjectConfig,
-  ProjectPollResult,
-  ProjectSubtask,
-} from '../types/Projects';
-import { ProjectDevTargetAccountPromptResponse } from '../types/Prompts';
+} from '../../buildAccount';
+import { hubspotAccountNamePrompt } from '../../prompts/accountNamePrompt';
+import { lib } from '../../../lang/en';
 import { FileResult } from 'tmp';
-import { Build } from '@hubspot/local-dev-lib/types/Build';
-import { getSandboxUsageLimits } from '@hubspot/local-dev-lib/api/sandboxHubs';
-
-const i18nKey = 'lib.localDev';
 
 // If the user passed in the --account flag, confirm they want to use that account as
 // their target account, otherwise exit
@@ -71,9 +77,9 @@ export async function confirmDefaultAccountIsTarget(
 ): Promise<void> {
   if (!accountConfig.name || !accountConfig.accountType) {
     logger.error(
-      i18n(`${i18nKey}.confirmDefaultAccountIsTarget.configError`, {
-        authCommand: uiCommandReference('hs auth'),
-      })
+      lib.localDevHelpers.confirmDefaultAccountIsTarget.configError(
+        uiCommandReference('hs auth')
+      )
     );
     process.exit(EXIT_CODES.ERROR);
   }
@@ -86,12 +92,9 @@ export async function confirmDefaultAccountIsTarget(
 
   if (!useDefaultAccount) {
     logger.log(
-      i18n(
-        `${i18nKey}.confirmDefaultAccountIsTarget.declineDefaultAccountExplanation`,
-        {
-          useCommand: uiCommandReference('hs accounts use'),
-          devCommand: uiCommandReference('hs project dev'),
-        }
+      lib.localDevHelpers.confirmDefaultAccountIsTarget.declineDefaultAccountExplanation(
+        uiCommandReference('hs accounts use'),
+        uiCommandReference('hs project dev')
       )
     );
     process.exit(EXIT_CODES.SUCCESS);
@@ -114,18 +117,18 @@ export async function checkIfDefaultAccountIsSupported(
     )
   ) {
     logger.error(
-      i18n(`${i18nKey}.checkIfDefaultAccountIsSupported.publicApp`, {
-        useCommand: uiCommandReference('hs accounts use'),
-        authCommand: uiCommandReference('hs auth'),
-      })
+      lib.localDevHelpers.checkIfDefaultAccountIsSupported.publicApp(
+        uiCommandReference('hs accounts use'),
+        uiCommandReference('hs auth')
+      )
     );
     process.exit(EXIT_CODES.SUCCESS);
   } else if (!hasPublicApps && isAppDeveloperAccount(accountConfig)) {
     logger.error(
-      i18n(`${i18nKey}.checkIfDefaultAccountIsSupported.privateApp`, {
-        useCommand: uiCommandReference('hs accounts use'),
-        authCommand: uiCommandReference('hs auth'),
-      })
+      lib.localDevHelpers.checkIfDefaultAccountIsSupported.privateApp(
+        uiCommandReference('hs accounts use'),
+        uiCommandReference('hs auth')
+      )
     );
     process.exit(EXIT_CODES.SUCCESS);
   }
@@ -137,15 +140,13 @@ export function checkIfParentAccountIsAuthed(accountConfig: CLIAccount): void {
     !getAccountConfig(accountConfig.parentAccountId)
   ) {
     logger.error(
-      i18n(`${i18nKey}.checkIfParentAccountIsAuthed.notAuthedError`, {
-        accountId: accountConfig.parentAccountId || '',
-        accountIdentifier: uiAccountDescription(
-          getAccountIdentifier(accountConfig)
-        ),
-        authCommand: uiCommandReference(
+      lib.localDevHelpers.checkIfParentAccountIsAuthed.notAuthedError(
+        uiCommandReference(
           `hs auth --account=${accountConfig.parentAccountId}`
         ),
-      })
+        accountConfig.parentAccountId || '',
+        uiAccountDescription(getAccountIdentifier(accountConfig))
+      )
     );
     process.exit(EXIT_CODES.SUCCESS);
   }
@@ -159,19 +160,19 @@ export function checkIfAccountFlagIsSupported(
   if (hasPublicApps) {
     if (!isDeveloperTestAccount(accountConfig)) {
       logger.error(
-        i18n(`${i18nKey}.validateAccountOption.invalidPublicAppAccount`, {
-          useCommand: uiCommandReference('hs accounts use'),
-          devCommand: uiCommandReference('hs project dev'),
-        })
+        lib.localDevHelpers.validateAccountOption.invalidPublicAppAccount(
+          uiCommandReference('hs accounts use'),
+          uiCommandReference('hs project dev')
+        )
       );
       process.exit(EXIT_CODES.SUCCESS);
     }
     checkIfParentAccountIsAuthed(accountConfig);
   } else if (isAppDeveloperAccount(accountConfig)) {
     logger.error(
-      i18n(`${i18nKey}.validateAccountOption.invalidPrivateAppAccount`, {
-        useCommand: uiCommandReference('hs accounts use'),
-      })
+      lib.localDevHelpers.validateAccountOption.invalidPrivateAppAccount(
+        uiCommandReference('hs accounts use')
+      )
     );
     process.exit(EXIT_CODES.SUCCESS);
   }
@@ -187,12 +188,11 @@ export async function suggestRecommendedNestedAccount(
   uiLine();
   if (hasPublicApps) {
     logger.log(
-      i18n(
-        `${i18nKey}.validateAccountOption.publicAppNonDeveloperTestAccountWarning`
-      )
+      lib.localDevHelpers.validateAccountOption
+        .publicAppNonDeveloperTestAccountWarning
     );
   } else {
-    logger.log(i18n(`${i18nKey}.validateAccountOption.nonSandboxWarning`));
+    logger.log(lib.localDevHelpers.validateAccountOption.nonSandboxWarning);
   }
   uiLine();
   logger.log();
@@ -218,18 +218,14 @@ export async function createSandboxForLocalDev(
     );
   } catch (err) {
     if (isMissingScopeError(err)) {
-      logger.error(
-        i18n('lib.sandbox.create.failure.scopes.message', {
-          accountName: accountConfig.name || accountId,
-        })
-      );
+      logger.error(lib.sandbox.create.developer.failure.scopes.message);
       const websiteOrigin = getHubSpotWebsiteOrigin(env);
       const url = `${websiteOrigin}/personal-access-key/${accountId}`;
       logger.info(
-        i18n('lib.sandbox.create.failure.scopes.instructions', {
-          accountName: accountConfig.name || accountId,
-          url,
-        })
+        lib.sandbox.create.developer.failure.scopes.instructions(
+          accountConfig.name || accountId,
+          url
+        )
       );
     } else {
       logError(err);
@@ -259,7 +255,7 @@ export async function createSandboxForLocalDev(
     const sandboxAccountConfig = getAccountConfig(result.sandbox.sandboxHubId);
 
     if (!sandboxAccountConfig) {
-      logger.error(i18n('lib.sandbox.create.failure.generic'));
+      logger.error(lib.sandbox.create.developer.failure.generic);
       process.exit(EXIT_CODES.ERROR);
     }
 
@@ -301,18 +297,14 @@ export async function createDeveloperTestAccountForLocalDev(
     }
   } catch (err) {
     if (isMissingScopeError(err)) {
-      logger.error(
-        i18n('lib.developerTestAccount.create.failure.scopes.message', {
-          accountName: accountConfig.name || accountId,
-        })
-      );
+      logger.error(lib.developerTestAccount.create.failure.scopes.message);
       const websiteOrigin = getHubSpotWebsiteOrigin(env);
       const url = `${websiteOrigin}/personal-access-key/${accountId}`;
       logger.info(
-        i18n('lib.developerTestAccount.create.failure.scopes.instructions', {
-          accountName: accountConfig.name || accountId,
-          url,
-        })
+        lib.developerTestAccount.create.failure.scopes.instructions(
+          accountConfig.name || accountId,
+          url
+        )
       );
     } else {
       logError(err);
@@ -355,12 +347,9 @@ export async function useExistingDevTestAccount(
   if (!useExistingDevTestAcct) {
     logger.log('');
     logger.log(
-      i18n(
-        `${i18nKey}.confirmDefaultAccountIsTarget.declineDefaultAccountExplanation`,
-        {
-          useCommand: uiCommandReference('hs accounts use'),
-          devCommand: uiCommandReference('hs project dev'),
-        }
+      lib.localDevHelpers.confirmDefaultAccountIsTarget.declineDefaultAccountExplanation(
+        uiCommandReference('hs accounts use'),
+        uiCommandReference('hs project dev')
       )
     );
     logger.log('');
@@ -372,10 +361,10 @@ export async function useExistingDevTestAccount(
     env
   );
   logger.success(
-    i18n(`lib.developerTestAccount.create.success.configFileUpdated`, {
-      accountName: devTestAcctConfigName,
-      authType: PERSONAL_ACCESS_KEY_AUTH_METHOD.name,
-    })
+    lib.developerTestAccount.create.success.configFileUpdated(
+      devTestAcctConfigName,
+      PERSONAL_ACCESS_KEY_AUTH_METHOD.name
+    )
   );
 }
 
@@ -390,34 +379,36 @@ export async function createNewProjectForLocalDev(
   let shouldCreateProject = shouldCreateWithoutConfirmation;
 
   if (!shouldCreateProject) {
-    const explanationString = i18n(
-      hasPublicApps
-        ? `${i18nKey}.createNewProjectForLocalDev.publicAppProjectMustExistExplanation`
-        : `${i18nKey}.createNewProjectForLocalDev.projectMustExistExplanation`,
-      {
-        accountIdentifier: uiAccountDescription(targetAccountId),
-        projectName: projectConfig.name,
-      }
+    const explanationLangFunction = hasPublicApps
+      ? lib.localDevHelpers.createNewProjectForLocalDev
+          .publicAppProjectMustExistExplanation
+      : lib.localDevHelpers.createNewProjectForLocalDev
+          .projectMustExistExplanation;
+
+    const explanationString = explanationLangFunction(
+      uiAccountDescription(targetAccountId),
+      projectConfig.name
     );
+
     logger.log();
     uiLine();
     logger.log(explanationString);
     uiLine();
 
     shouldCreateProject = await confirmPrompt(
-      i18n(`${i18nKey}.createNewProjectForLocalDev.createProject`, {
-        accountIdentifier: uiAccountDescription(targetAccountId),
-        projectName: projectConfig.name,
-      })
+      lib.localDevHelpers.createNewProjectForLocalDev.createProject(
+        projectConfig.name,
+        uiAccountDescription(targetAccountId)
+      )
     );
   }
 
   if (shouldCreateProject) {
     SpinniesManager.add('createProject', {
-      text: i18n(`${i18nKey}.createNewProjectForLocalDev.creatingProject`, {
-        accountIdentifier: uiAccountDescription(targetAccountId),
-        projectName: projectConfig.name,
-      }),
+      text: lib.localDevHelpers.createNewProjectForLocalDev.creatingProject(
+        projectConfig.name,
+        uiAccountDescription(targetAccountId)
+      ),
     });
 
     try {
@@ -426,17 +417,17 @@ export async function createNewProjectForLocalDev(
         projectConfig.name
       );
       SpinniesManager.succeed('createProject', {
-        text: i18n(`${i18nKey}.createNewProjectForLocalDev.createdProject`, {
-          accountIdentifier: uiAccountDescription(targetAccountId),
-          projectName: projectConfig.name,
-        }),
+        text: lib.localDevHelpers.createNewProjectForLocalDev.createdProject(
+          projectConfig.name,
+          uiAccountDescription(targetAccountId)
+        ),
         succeedColor: 'white',
       });
       return project;
     } catch (err) {
       SpinniesManager.fail('createProject');
       logger.log(
-        i18n(`${i18nKey}.createNewProjectForLocalDev.failedToCreateProject`)
+        lib.localDevHelpers.createNewProjectForLocalDev.failedToCreateProject
       );
       process.exit(EXIT_CODES.ERROR);
     }
@@ -444,7 +435,7 @@ export async function createNewProjectForLocalDev(
     // We cannot continue if the project does not exist in the target account
     logger.log();
     logger.log(
-      i18n(`${i18nKey}.createNewProjectForLocalDev.choseNotToCreateProject`)
+      lib.localDevHelpers.createNewProjectForLocalDev.choseNotToCreateProject
     );
     process.exit(EXIT_CODES.SUCCESS);
   }
@@ -458,9 +449,9 @@ function projectUploadCallback(
 ): Promise<ProjectPollResult> {
   if (!buildId) {
     logger.error(
-      i18n(`${i18nKey}.createInitialBuildForNewProject.initialUploadMessage`, {
-        uploadCommand: uiCommandReference('hs project upload'),
-      })
+      lib.localDevHelpers.createInitialBuildForNewProject.genericError(
+        uiCommandReference('hs project upload')
+      )
     );
     process.exit(EXIT_CODES.ERROR);
   }
@@ -488,7 +479,7 @@ export async function createInitialBuildForNewProject(
       projectConfig,
       projectDir,
       projectUploadCallback,
-      i18n(`${i18nKey}.createInitialBuildForNewProject.initialUploadMessage`),
+      lib.localDevHelpers.createInitialBuildForNewProject.initialUploadMessage,
       sendIr
     );
 
@@ -500,7 +491,7 @@ export async function createInitialBuildForNewProject(
     ) {
       logger.log();
       logger.error(
-        i18n(`${i18nKey}.createInitialBuildForNewProject.projectLockedError`)
+        lib.localDevHelpers.createInitialBuildForNewProject.projectLockedError
       );
       logger.log();
     } else {
